==> S0.MART.Rmd <==
---
title: "MART Flexible Expenditure"
author: "Lutfor Rahman"
date: "2023-09-28"
output:
  cleanrmd::html_document_clean:
    theme: superstylin
    toc: true
    toc_float: true
---


```{r setup, include=FALSE, echo=FALSE, message=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
library(tidyverse)
library(foreign)
library(srvyr)

rm(list=ls())
```

#Expenditure Flexible from MART 1. Utilities; 2. Food; 3. Personal 4. Leisure 
#by Region, Household Type and Disability
```{r}
MART<-readRDS("MART.Dash.rds")
names(MART)

MART<-MART%>% mutate(Household.Type.Child=case_when(Demographic.Household.Long=="Single Person"  ~"Single Person",
                                                   Demographic.Household.Long=="Couple"  ~"Couple",
                                                   
                                                   (Demographic.Household.Long=="Couple With Dependent Children"|
                                                   Demographic.Household.Long=="Couple With Non-Dependent Children") & 
                                                   (Demographic.Children.Short=="One") ~"Couple With 1 child",
                                                  
                                                   (Demographic.Household.Long=="Couple With Dependent Children"|
                                                    Demographic.Household.Long=="Couple With Non-Dependent Children") & 
                                                   (Demographic.Children.Short=="Two") ~"Couple With 2 child",
                                                  
                                                   (Demographic.Household.Long=="Couple With Dependent Children"|
                                                    Demographic.Household.Long=="Couple With Non-Dependent Children") & 
                                                   (Demographic.Children.Short=="Three") ~"Couple With 3 child",
                                                 
                                                   (Demographic.Household.Long=="Couple With Dependent Children" |
                                                    Demographic.Household.Long=="Couple With Non-Dependent Children") & 
                                                   (Demographic.Children.Short=="Four"|
                                                    Demographic.Children.Short=="Greater than 4") ~"Couple With 4 and more child",
                                                  
                                                   (Demographic.Household.Long=="Single Person With Dependent Children"|
                                                    Demographic.Household.Long=="Single Person With Non-Dependent Children") & 
                                                   (Demographic.Children.Short=="One") ~"Single With 1 child",
                                                  
                                                   (Demographic.Household.Long=="Single Person With Dependent Children"|
                                                    Demographic.Household.Long=="Single Person With Non-Dependent Children") & 
                                                   (Demographic.Children.Short=="Two"|
                                                    Demographic.Children.Short=="Three"| 
                                                    Demographic.Children.Short=="Four" | 
                                                    Demographic.Children.Short=="Greater than 4") ~"Single With 2 or more child",
                               
                            
                                                  (Demographic.Household.Long=="Other adults and Dependent children"|
                                                   Demographic.Household.Long=="Other adults and Non-Dependent children") & 
                                                  (Demographic.Children.Short=="One"|
                                                   Demographic.Children.Short=="Two"|
                                                   Demographic.Children.Short=="Three"| 
                                                   Demographic.Children.Short=="Four" | 
                                                   Demographic.Children.Short=="Greater than 4") ~"Adults With 1 or more child",
                                 
                                                   Demographic.Household.Long=="Other Adults Only"~"Adults Only",
                               
                                                  TRUE~ "Other"))
```



```{r}
MART_1<-MART %>% select(Demographic.Region, 
                          Financial.Year, 
                          Demographic.Health, 
                          Demographic.Household, 
                          Demographic.Household.Long,  
                          Demographic.Children.Short, 
                          Household.Type.Child,
                          Income.Total,
                          Income.Attendance.Allowance, 
                          Income.DLA, 
                          Income.Disability.benefits,
                          Income.PIP,
                          Expenditure.Total, 
                        
                          Expenditure.Communications.and.Leisure.Total, 
                          Expenditure.Mobile, 
                          Expenditure.Broadband,
                          Expenditure.Food.and.Housekeeping.Total, 
                          Expenditure.Utilities.Total, 
                          Expenditure.Personal.Total,
                          Expenditure.Clothing.and.Footwear)
```


```{r}
MART_1<-MART_1 %>% 
    mutate(Health.Condition=case_when(Demographic.Health=="Disabled"|
                                      Demographic.Health=="Long-Term Health Condition" ~"disabled",
                                      Demographic.Health=="None"~ "None",
                                      TRUE~ "other/not known")) %>% 
    mutate(AA=case_when(Income.Attendance.Allowance>0~ "1", TRUE~"0")) %>% 
    mutate(DLA=case_when(Income.DLA>0 ~"1", TRUE~"0")) %>% 
    mutate(PIP=case_when(Income.PIP>0 ~"1", TRUE~"0")) %>%
    mutate(DB=case_when(Income.Disability.benefits>0 ~"1", TRUE~"0")) 

MART_1<-MART_1 %>% mutate(Disability=case_when(AA==1|DLA==1|PIP==1|DB==1 ~ "Disability.Benefits",
                                      (Health.Condition=="disabled" & (AA==0 | DLA==0 | PIP==0 | DB==0))~"Disability.No.Benefits",
                                             
                                             TRUE~"None")) 

#Disability variable being created by disability benefits (if clients have either one of these AA|DLA|PIP|DB) or disability no benefits 
#diababled but not recepient of any above, last category is none
```


```{r}

Expenditure.Flexible.MART.21<-MART_1 %>%
  dplyr::group_by(Financial.Year, Household.Type.Child, Demographic.Region, Disability) %>%
  mutate(Expenditure.Leisure.MART=Expenditure.Communications.and.Leisure.Total-(Expenditure.Mobile+Expenditure.Broadband),
         Expenditure.Mobile.HomePhone.MART=(Expenditure.Mobile+Expenditure.Broadband))%>%
  summarise(Sample.MART=n(),
            Expenditure.Leisure.21.MART=(median(Expenditure.Leisure.MART, na.rm=TRUE))*12/52,
            Expenditure.Mobile.HomePhone.21.MART=(median(Expenditure.Mobile.HomePhone.MART, na.rm=TRUE))*12/52,
            Expenditure.Food.House.Keeping.21.MART=(median(Expenditure.Food.and.Housekeeping.Total, na.rm = TRUE))*12/52,
            Expenditure.Personal.Cost.21.MART=(median(Expenditure.Personal.Total, na.rm=TRUE))*12/52,
            Expenditure.Utilities.21.MART=(median(Expenditure.Utilities.Total, na.rm=TRUE))*12/52,
            Derived.Flexible.Expenditure.Total.21.MART=sum(Expenditure.Food.House.Keeping.21.MART,
                                                        Expenditure.Personal.Cost.21.MART,
                                                        Expenditure.Leisure.21.MART, 
                                                        Expenditure.Utilities.21.MART)) %>% 
  #filter(Household.Type!="Other/prefer not to say/unknown") %>% 
  #filter(Region!="Other/not specified") %>% 
  filter(Financial.Year=="2020-21") 

saveRDS(Expenditure.Flexible.MART.21, "Expenditure.Flexible.MART.21.rds")


##

Expenditure.Flexible.MART.22<-MART_1 %>%
 dplyr::group_by(Financial.Year,  Household.Type.Child, Demographic.Region, Disability) %>%
 mutate(Expenditure.Leisure.MART=Expenditure.Communications.and.Leisure.Total-(Expenditure.Mobile+Expenditure.Broadband),
         Expenditure.Mobile.HomePhone.MART=(Expenditure.Mobile+Expenditure.Broadband))%>%
  summarise(Sample.MART=n(),
            Expenditure.Leisure.22.MART=(median(Expenditure.Leisure.MART, na.rm=TRUE))*12/52,
            Expenditure.Mobile.HomePhone.22.MART=(median(Expenditure.Mobile.HomePhone.MART, na.rm=TRUE))*12/52,
            Expenditure.Food.House.Keeping.22.MART=(median(Expenditure.Food.and.Housekeeping.Total, na.rm = TRUE))*12/52,
            Expenditure.Personal.Cost.22.MART=(median(Expenditure.Personal.Total, na.rm=TRUE))*12/52,
            Expenditure.Utilities.22.MART=(median(Expenditure.Utilities.Total, na.rm=TRUE))*12/52,
            Derived.Flexible.Expenditure.Total.22.MART=sum(Expenditure.Food.House.Keeping.22.MART,
                                                        Expenditure.Personal.Cost.22.MART,
                                                        Expenditure.Leisure.22.MART, 
                                                        Expenditure.Utilities.22.MART)) %>% 
  #filter(Household.Type!="Other/prefer not to say/unknown") %>% 
  #filter(Region!="Other/not specified") %>% 
  filter(Financial.Year=="2021-22") 

saveRDS(Expenditure.Flexible.MART.22, "Expenditure.Flexible.MART.22.rds")

Expenditure.Flexible.MART.23<-MART_1 %>%
 dplyr::group_by(Financial.Year, Household.Type.Child, Demographic.Region, Disability) %>%
  mutate(Expenditure.Leisure.MART=Expenditure.Communications.and.Leisure.Total-(Expenditure.Mobile+Expenditure.Broadband),
         Expenditure.Mobile.HomePhone.MART=(Expenditure.Mobile+Expenditure.Broadband))%>%
  summarise(Sample.MART=n(),
            Expenditure.Leisure.23.MART=(median(Expenditure.Leisure.MART, na.rm=TRUE))*12/52,
            Expenditure.Mobile.HomePhone.23.MART=(median(Expenditure.Mobile.HomePhone.MART, na.rm=TRUE))*12/52,
            Expenditure.Food.House.Keeping.23.MART=(median(Expenditure.Food.and.Housekeeping.Total, na.rm = TRUE))*12/52,
            Expenditure.Personal.Cost.23.MART=(median(Expenditure.Personal.Total, na.rm=TRUE))*12/52,
            Expenditure.Utilities.23.MART=(median(Expenditure.Utilities.Total, na.rm=TRUE))*12/52,
            Derived.Flexible.Expenditure.Total.23.MART=sum(Expenditure.Food.House.Keeping.23.MART,
                                                        Expenditure.Personal.Cost.23.MART,
                                                        Expenditure.Leisure.23.MART, 
                                                        Expenditure.Utilities.23.MART)) %>% 
  #filter(Household.Type!="Other/prefer not to say/unknown") %>% 
  #filter(Region!="Other/not specified") %>% 
  filter(Financial.Year=="2022-23") 

saveRDS(Expenditure.Flexible.MART.23, "Expenditure.Flexible.MART.23.rds")

###
Expenditure.Flexible.MART.24<-MART_1 %>%
 dplyr::group_by(Financial.Year, Household.Type.Child, Demographic.Region, Disability) %>%
  mutate(Expenditure.Leisure.MART=Expenditure.Communications.and.Leisure.Total-(Expenditure.Mobile+Expenditure.Broadband),
         Expenditure.Mobile.HomePhone.MART=(Expenditure.Mobile+Expenditure.Broadband))%>%
  summarise(Sample.MART=n(),
            Expenditure.Leisure.24.MART=(median(Expenditure.Leisure.MART, na.rm=TRUE))*12/52,
            Expenditure.Mobile.HomePhone.24.MART=(median(Expenditure.Mobile.HomePhone.MART, na.rm=TRUE))*12/52,
            Expenditure.Food.House.Keeping.24.MART=(median(Expenditure.Food.and.Housekeeping.Total, na.rm = TRUE))*12/52,
            Expenditure.Personal.Cost.24.MART=(median(Expenditure.Personal.Total, na.rm=TRUE))*12/52,
            Expenditure.Utilities.24.MART=(median(Expenditure.Utilities.Total, na.rm=TRUE))*12/52,
            Derived.Flexible.Expenditure.Total.24.MART=sum(Expenditure.Food.House.Keeping.24.MART,
                                                        Expenditure.Personal.Cost.24.MART,
                                                        Expenditure.Leisure.24.MART, 
                                                        Expenditure.Utilities.24.MART)) %>% 
  #filter(Household.Type!="Other/prefer not to say/unknown") %>% 
  #filter(Region!="Other/not specified") %>% 
  filter(Financial.Year=="2023-24") 

saveRDS(Expenditure.Flexible.MART.24, "Expenditure.Flexible.MART.24.rds")

```

```{r}
NI<-Expenditure.Flexible.MART.21 %>% filter(Demographic.Region=="London") 
Scotland<-Expenditure.Flexible.MART.21 %>% filter(Demographic.Region=="North West") 
  
NI$Demographic.Region <- recode_factor(NI$Demographic.Region,  "London"="Northern Ireland")
Scotland$Demographic.Region<-recode_factor(Scotland$Demographic.Region,  "North West"="Scotland")
```
