==> S2.LCFS.20_21.Variables.Rmd <==
---
title: "LCFS Variables 2020-21"
author: "Lutfor Rahman"
date: "2023-09-28"
output:
  cleanrmd::html_document_clean:
    theme: superstylin
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
```

#Pulling Benefits variables
```{r}

#Loading LCFS raw person level data for 2020-21
# LCFS_21_person<-read.spss("lcfs_2020_rawper_ukanon_final.sav", #FY 2020-21
#                         use.value.labels = TRUE, 
#                         use.missings = FALSE, 
#                         to.data.frame = TRUE)
# 
# saveRDS(LCFS_21_person, "LCFS_21_person.rds")

LCFS_21_person<-readRDS("LCFS_21_person.rds")

#Subset of variables which have benefit elements

LCFS_21_benefits <- LCFS_21_person %>%
  select(case, Person, Nolwm, DVUC, DVHB, DVWTC, DVCTC,DVJSACON, DVJSAIB, DVJSAComb, DVESA,DVPIPCar,
         DVPIPMob, DVCB, DVPC, DVIS, DVExPay, DVCA) %>%
  pivot_longer(cols = c(DVUC, DVHB, DVWTC, DVCTC,DVJSACON, DVJSAIB, DVJSAComb, DVESA,DVPIPCar,
         DVPIPMob, DVCB, DVPC, DVIS, DVExPay, DVCA), 
               names_to = "Demographic",
               values_to = "Income") %>%
  mutate(Demographic = case_when(Income > 0 ~ Demographic, TRUE ~ "No" ))


LCFS_21_benefits_result <- LCFS_21_benefits %>%
  filter(Demographic != "No") %>%
  group_by(case, Demographic) %>%
  summarise(n = n(),
            Income = sum(Income)) 


LCFS_21_benefits_result_wide <- LCFS_21_benefits_result %>%
  spread(key = Demographic, value = Income, fill = 0)


LCFS_21_benefits_result_wide<-LCFS_21_benefits_result_wide %>% 
  rename("Income.UC.21"=DVUC,
         "Income.Housing.Benefit.21"=DVHB,
         "Income.WTC.21"=DVWTC, 
         "Income.Child.Tax.Credit.21"= DVCTC,
         "Income.JSA.Contribution.Based.21"=DVJSACON, 
         "Income.JSA.Income.Based.21"=DVJSAIB, 
         "Income.JSA.Combined.21"=DVJSAComb, 
         "Income.ESA.21"=DVESA,
         "Income.PIP.Car.21"= DVPIPCar,
         "Income.PIP.Mob.21"=DVPIPMob, 
         "Income.Child.Benefit.21"= DVCB,
         "Income.Pension.Credit.21"=DVPC, 
         "Income.Income.Support.21"=DVIS,
         "Income.ExtPaym.Housing.Benefit.21"=DVExPay,
         "Income.Carer.Allowance.21"=DVCA)
```


```{r}
LCFS_21_benefits_result_total <- LCFS_21_benefits_result_wide %>%
  group_by(case) %>%
  summarise(across(starts_with("Income"), sum, na.rm = TRUE))
```




```{r}
# Inflation in 2021-22 of all types of benefits is 0.5% and 2022-23 is 3.1%
LCFS_21_benefits_result_total$benefit.inflation.22 <-1.005
LCFS_21_benefits_result_total$benefit.inflation.23 <-1.031
LCFS_21_benefits_result_total$benefit.inflation.24 <-1.101


LCFS_21_benefits_result_total<-LCFS_21_benefits_result_total%>% 
  mutate(Income.UC.22=Income.UC.21*benefit.inflation.22,
         Income.Housing.Benefit.22=Income.Housing.Benefit.21*benefit.inflation.22,
         Income.WTC.22=Income.WTC.21*benefit.inflation.22, 
         Income.Child.Tax.Credit.22=Income.Child.Tax.Credit.21*benefit.inflation.22 ,
         Income.JSA.Contribution.Based.22=Income.JSA.Contribution.Based.21*benefit.inflation.22, 
         Income.JSA.Income.Based.22=Income.JSA.Income.Based.21*benefit.inflation.22, 
         Income.JSA.Combined.22=Income.JSA.Combined.21*benefit.inflation.22, 
         Income.ESA.22=Income.ESA.21*benefit.inflation.22,
         Income.PIP.Car.22=Income.PIP.Car.21*benefit.inflation.22 ,
         Income.PIP.Mob.22=Income.PIP.Mob.21*benefit.inflation.22, 
         Income.Child.Benefit.22=Income.Child.Benefit.21*benefit.inflation.22,
         Income.Pension.Credit.22=Income.Pension.Credit.21*benefit.inflation.22, 
         Income.Income.Support.22=Income.Income.Support.21*benefit.inflation.22,
         Income.ExtPaym.Housing.Benefit.22=Income.ExtPaym.Housing.Benefit.21*benefit.inflation.22,
         Income.Carer.Allowance.22=Income.Carer.Allowance.21*benefit.inflation.22)


LCFS_21_benefits_result_total<-LCFS_21_benefits_result_total%>% 
  mutate(Income.UC.23=Income.UC.22*benefit.inflation.23,
         Income.Housing.Benefit.23=Income.Housing.Benefit.22*benefit.inflation.23,
         Income.WTC.23=Income.WTC.22*benefit.inflation.23, 
         Income.Child.Tax.Credit.23=Income.Child.Tax.Credit.22*benefit.inflation.23 ,
         Income.JSA.Contribution.Based.23=Income.JSA.Contribution.Based.22*benefit.inflation.23, 
         Income.JSA.Income.Based.23=Income.JSA.Income.Based.22*benefit.inflation.23, 
         Income.JSA.Combined.23=Income.JSA.Combined.22*benefit.inflation.23, 
         Income.ESA.23=Income.ESA.22*benefit.inflation.23,
         Income.PIP.Car.23=Income.PIP.Car.22*benefit.inflation.23 ,
         Income.PIP.Mob.23=Income.PIP.Mob.22*benefit.inflation.23, 
         Income.Child.Benefit.23=Income.Child.Benefit.22*benefit.inflation.23,
         Income.Pension.Credit.23=Income.Pension.Credit.22*benefit.inflation.23, 
         Income.Income.Support.23=Income.Income.Support.22*benefit.inflation.23,
         Income.ExtPaym.Housing.Benefit.23=Income.ExtPaym.Housing.Benefit.22*benefit.inflation.23,
         Income.Carer.Allowance.23=Income.Carer.Allowance.22*benefit.inflation.23)


LCFS_21_benefits_result_total<-LCFS_21_benefits_result_total%>% 
  mutate(Income.UC.24=Income.UC.23*benefit.inflation.24,
         Income.Housing.Benefit.24=Income.Housing.Benefit.23*benefit.inflation.24,
         Income.WTC.24=Income.WTC.23*benefit.inflation.24, 
         Income.Child.Tax.Credit.24=Income.Child.Tax.Credit.23*benefit.inflation.24,
         Income.JSA.Contribution.Based.24=Income.JSA.Contribution.Based.23*benefit.inflation.24, 
         Income.JSA.Income.Based.24=Income.JSA.Income.Based.23*benefit.inflation.24, 
         Income.JSA.Combined.24=Income.JSA.Combined.23*benefit.inflation.24, 
         Income.ESA.24=Income.ESA.23*benefit.inflation.24,
         Income.PIP.Car.24=Income.PIP.Car.23*benefit.inflation.24,
         Income.PIP.Mob.24=Income.PIP.Mob.23*benefit.inflation.24, 
         Income.Child.Benefit.24=Income.Child.Benefit.23*benefit.inflation.24,
         Income.Pension.Credit.24=Income.Pension.Credit.23*benefit.inflation.24, 
         Income.Income.Support.24=Income.Income.Support.23*benefit.inflation.24,
         Income.ExtPaym.Housing.Benefit.24=Income.ExtPaym.Housing.Benefit.23*benefit.inflation.24,
         Income.Carer.Allowance.24=Income.Carer.Allowance.23*benefit.inflation.24)
```

# Taper off benefits
```{r}

```



```{r}
LCFS_benefits_21<-LCFS_21_benefits_result_total%>% mutate(Year="2020-21")
LCFS_benefits_22P<-LCFS_21_benefits_result_total%>% mutate(Year="2021-22P")
LCFS_benefits_23P<-LCFS_21_benefits_result_total%>% mutate(Year="2022-23P")
LCFS_benefits_24P<-LCFS_21_benefits_result_total%>% mutate(Year="2023-24P")

LCFS_21_benefits_all<- rbind(LCFS_benefits_21, LCFS_benefits_22P,LCFS_benefits_23P, LCFS_benefits_24P)
```

```{r}
LCFS_21_benefits_all <- LCFS_21_benefits_all %>%
  mutate(Income.UC = case_when(Year == "2020-21" ~ Income.UC.21,
                               Year == "2021-22P" ~ Income.UC.22,
                               Year == "2022-23P" ~ Income.UC.23,
                               Year == "2023-24P" ~ Income.UC.24)) %>% 
  
  mutate(Income.Housing.Benefit=case_when(Year == "2020-21" ~ Income.Housing.Benefit.21,
                                          Year == "2021-22P" ~ Income.Housing.Benefit.22,
                                          Year == "2022-23P" ~ Income.Housing.Benefit.23,
                                          Year == "2023-24P" ~ Income.Housing.Benefit.24))%>%
  
  mutate(Income.WTC=case_when(Year == "2020-21" ~ Income.WTC.21,
                              Year == "2021-22P" ~ Income.WTC.22,
                              Year == "2022-23P" ~ Income.WTC.23,
                              Year == "2023-24P" ~ Income.WTC.24))%>%

  
  mutate(Income.Child.Tax.Credit=case_when(Year == "2020-21" ~ Income.Child.Tax.Credit.21,
                                           Year == "2021-22P" ~ Income.Child.Tax.Credit.22,
                                           Year == "2022-23P" ~ Income.Child.Tax.Credit.23,
                                           Year == "2023-24P" ~ Income.Child.Tax.Credit.24)) %>% 
 
  mutate(Income.JSA.Contribution.Based=case_when(Year == "2020-21" ~ Income.JSA.Contribution.Based.21,
                                                 Year == "2021-22P" ~ Income.JSA.Contribution.Based.22,
                                                 Year == "2022-23P" ~ Income.JSA.Contribution.Based.23,
                                                 Year == "2023-24P" ~ Income.JSA.Contribution.Based.24))   %>% 
 
  mutate(Income.JSA.Income.Based=case_when(Year == "2020-21" ~ Income.JSA.Income.Based.21,
                                           Year == "2021-22P" ~ Income.JSA.Income.Based.22,
                                           Year == "2022-23P" ~ Income.JSA.Income.Based.23,
                                           Year == "2023-24P" ~ Income.JSA.Income.Based.24))   %>% 
 
  mutate(Income.JSA.Combined=case_when(Year == "2020-21" ~ Income.JSA.Combined.21,
                                       Year == "2021-22P" ~ Income.JSA.Combined.22,
                                       Year == "2022-23P" ~ Income.JSA.Combined.23,
                                       Year == "2023-24P" ~ Income.JSA.Combined.24))  %>% 
 
  mutate(Income.ESA=case_when(Year == "2020-21" ~ Income.ESA.21,
                              Year == "2021-22P" ~ Income.ESA.22,
                              Year == "2022-23P" ~ Income.ESA.23,
                              Year == "2023-24P" ~ Income.ESA.24)) %>% 
  
  mutate(Income.PIP.Car=case_when(Year == "2020-21" ~ Income.PIP.Car.21,
                                    Year == "2021-22P" ~ Income.PIP.Car.22,
                                    Year == "2022-23P" ~ Income.PIP.Car.23,
                                    Year == "2023-24P" ~ Income.PIP.Car.24)) %>% 
  
  mutate(Income.PIP.Mob=case_when(Year == "2020-21" ~ Income.PIP.Mob.21,
                                   Year == "2021-22P" ~ Income.PIP.Mob.22,
                                   Year == "2022-23P" ~ Income.PIP.Mob.23,
                                   Year == "2023-24P" ~ Income.PIP.Mob.24)) %>% 
  
  mutate(Income.Child.Benefit=case_when(Year == "2020-21" ~ Income.Child.Benefit.21,
                                          Year == "2021-22P" ~ Income.Child.Benefit.22,
                                          Year == "2022-23P" ~ Income.Child.Benefit.23,
                                          Year == "2023-24P" ~ Income.Child.Benefit.24)) %>%  
  
  mutate(Income.Pension.Credit=case_when(Year == "2020-21" ~ Income.Pension.Credit.21,
                                          Year == "2021-22P" ~ Income.Pension.Credit.22,
                                          Year == "2022-23P" ~ Income.Pension.Credit.23,
                                          Year == "2023-24P" ~ Income.Pension.Credit.24))%>% 
  
  mutate(Income.Income.Support=case_when(Year == "2020-21" ~ Income.Income.Support.21,
                                          Year == "2021-22P" ~ Income.Income.Support.22,
                                          Year == "2022-23P" ~ Income.Income.Support.23,
                                          Year == "2023-24P" ~ Income.Income.Support.24))%>% 
  
  mutate(Income.ExtPaym.Housing.Benefit=case_when(Year == "2020-21" ~ Income.ExtPaym.Housing.Benefit.21,
                                          Year == "2021-22P" ~ Income.ExtPaym.Housing.Benefit.22,
                                          Year == "2022-23P" ~ Income.ExtPaym.Housing.Benefit.23,
                                          Year == "2023-24P" ~ Income.ExtPaym.Housing.Benefit.24))%>% 
  
  mutate(Income.Carer.Allowance=case_when(Year == "2020-21" ~ Income.Carer.Allowance.21,
                                          Year == "2021-22P" ~ Income.Carer.Allowance.22,
                                          Year == "2022-23P" ~ Income.Carer.Allowance.23,
                                          Year == "2023-24P" ~ Income.Carer.Allowance.24))
    
```


```{r}
LCFS_21_benefits_all<-LCFS_21_benefits_all %>% select("case", 
                                                "Year", 
                                                "Income.UC",
         "Income.Housing.Benefit",
         "Income.WTC", 
         "Income.Child.Tax.Credit",
         "Income.JSA.Contribution.Based", 
         "Income.JSA.Income.Based", 
         "Income.JSA.Combined", 
         "Income.ESA",
         "Income.PIP.Car",
         "Income.PIP.Mob", 
         "Income.Child.Benefit",
         "Income.Pension.Credit",
         "Income.Income.Support",
         "Income.ExtPaym.Housing.Benefit",
         "Income.Carer.Allowance") 
saveRDS(LCFS_21_benefits_all, "LCFS_21_benefits_all.rds")
```

#Number of children in different age group
```{r}
LCFS_21_child<- LCFS_21_person %>%select(case,Person,DVBenU1,HRPID,HRPPrID, Sex, 
                                         HasChd, HasDep, HasNDep,NChild, NDepC, 
                                         NNDepC, NBaby, NCUnd5, NC5to9, NC1015, NCU16, NC1618) %>% 
                                filter(HasChd=="Yes")

LCFS_21_child1<-LCFS_21_child %>% filter(DVBenU1>1)

LCFS_21_child_summary <- LCFS_21_child %>% 
  group_by(case) %>% 
  summarise(
    Mean_NChild = round(mean(NChild, na.rm = TRUE), 1),
    Mean_NDepC = round(mean(NDepC, na.rm = TRUE), 1),
    Mean_NNDepC = round(mean(NNDepC, na.rm = TRUE), 1),
    Mean_NBaby = round(mean(NBaby, na.rm = TRUE), 1),
    Mean_NCUnd5 = round(mean(NCUnd5, na.rm = TRUE), 1),
    Mean_NC5to9 = round(mean(NC5to9, na.rm = TRUE), 1),
    Mean_NC1015 = round(mean(NC1015, na.rm = TRUE), 1),
    Mean_NCU16 = round(mean(NCU16, na.rm = TRUE), 1),
    Mean_NC1618 = round(mean(NC1618, na.rm = TRUE), 1)
  ) %>%    
  rename("Demographic.Number.of.Child.LCFS"=Mean_NChild, 
         "Demographic.Number.of.Dependent.Child.LCFS" =Mean_NDepC,
         "Demographic.Number.of.Non.Dependent.Child.LCFS"=Mean_NNDepC, 
          "Demographic.Number.of.Child.<1.LCFS"=Mean_NBaby,
          "Demographic.Number.of.Child.<5.LCFS"=Mean_NCUnd5, 
          "Demographic.Number.of.Child.5-9.LCFS"=Mean_NC5to9,
         "Demographic.Number.of.Child.10-15.LCFS"=Mean_NC1015,
         "Demographic.Number.of.Child.<16.LCFS"=Mean_NCU16,
         "Demographic.Number.of.Child.16-18.LCFS"=Mean_NC1618
         )

saveRDS(LCFS_21_child_summary, "LCFS_21_child_summary.rds")   

```




```{r}

# Work_Hours_21<-LCFS_21_person %>% select(case, Person, DVUsHr,DVTotHrU, Totus1, UsuHr, HRPID, HRPPrID, R01, DVILO3a) %>% 
#   rename("Demographic.Total.Usual.Hours.LCFS"=DVUsHr, 
#           "Demographic.Total.Usual.Hours.With.Overtime.LCFS"=DVTotHrU,
#          "Demographic.Total.Usual.Hours.Main.Job.LCFS" =Totus1,
#           "Demographic.Total.Usual.Hours.Main.Job.Without.Break.LCFS"=UsuHr, 
#           "Demographic.Household.Reference.ID.LCFS"=HRPID,
#           "Demographic.Household.Reference.Partner.LCFS"=HRPPrID, 
#           "Demographic.Relationship.LCFS"=R01,
#          "Demographic.Employment.Status.LCFS"=DVILO3a
#          )
# 
# 
# hrp<-Work_Hours_21 %>% filter(Demographic.Household.Reference.ID.LCFS=="HRP") %>% filter(!duplicated(case))
# 
# hrp_partner<-Work_Hours_21 %>% filter(Demographic.Household.Reference.Partner.LCFS=="Partner of HRP") %>% filter(!duplicated(case))
# df_hrp_partner<-hrp %>% left_join(hrp_partner, by=c("case"="case"))
# 
# names(df_hrp_partner)
# 
# LCFS_21_Work_Hours<-df_hrp_partner %>% select(case, 
#                                            Demographic.Household.Reference.ID.LCFS.x,
#                                            Demographic.Employment.Status.LCFS.x, 
#                                            Demographic.Total.Usual.Hours.LCFS.x, 
#                                            Demographic.Household.Reference.Partner.LCFS.y,
#                                            Demographic.Employment.Status.LCFS.y,
#                                            Demographic.Total.Usual.Hours.LCFS.y 
#                                            ) %>% 
#                                   rename("Demographic.Household.Reference.ID.LCFS"=Demographic.Household.Reference.ID.LCFS.x,
#                                          "Demographic.HRP. Employment.Status.LCFS" = Demographic.Employment.Status.LCFS.x, 
#                                          "Demographic.HRP.Total.Usual.Hours.LCFS"= Demographic.Total.Usual.Hours.LCFS.x, 
#                                          "Demographic.Household.Reference.Partner.LCFS" =Demographic.Household.Reference.Partner.LCFS.y,
#                                          "Demographic.HRP.Partner.Employment.Status.LCFS"=Demographic.Employment.Status.LCFS.y,
#                                          "Demographic.Partner.Total.Usual.Hours.LCFS" =Demographic.Total.Usual.Hours.LCFS.y 
#                                            )
#   
# 
# saveRDS(LCFS_21_Work_Hours, "LCFS_21_Work_Hours.rds")
```

```{r}
#Child Gender
#From households level data
LCFS_21<-readRDS("LCFS_21.rds")

LCFS_Child_21_Gender<-LCFS_21 %>% select(case, A020,A021,A022, A030,A031, A032) %>% 
  mutate(Demographic.Child.Gender=case_when((A020>0|A021>0|A022>0) &
                                            (A030>0|A031>0| A032>0)~ "Male & Female",
                                            A020>0|A021>0|A022>0~ "Male",
                                            A030>0|A031>0| A032>0~ "Female",
                                            (
                                             TRUE~"Other"
                                            ))) %>% select(case, Demographic.Child.Gender)


saveRDS(LCFS_Child_21_Gender, "LCFS_Child_21_Gender.rds")

```

#HRP and Partner wages and hours work

```{r}
#To understand HRP total hours and total wages
LCFS_21_dv_person<-read.spss("lcfs_2020_dvper_ukanon202021.sav", #FY 2021-22
                             use.value.labels = TRUE,
                             use.missings = FALSE,
                             to.data.frame = TRUE)
saveRDS(LCFS_21_dv_person, "LCFS_21_dv_person.rds")
 


LCFS_21_dv_person<-readRDS("LCFS_21_dv_person.rds")

LCFS_21_dv_person1<-LCFS_21_dv_person %>% select(case, Person, A055, A002,
A003,A0031, A220,P004p) 

LCFS_21_dv_HRP<-LCFS_21_dv_person1 %>% select(case,A055, 
A003,A220,P004p) %>% filter(A003=="HRP")

LCFS_21_dv_HRP_Partner<-LCFS_21_dv_person1 %>% select(case, 
A0031,A220,P004p) %>% filter(A0031=="Partner of HRP")


LCFS_21_Work<-LCFS_21_dv_HRP%>% 
  left_join(LCFS_21_dv_HRP_Partner, by=c("case"="case")) %>% 
  rename("Sample.Month"=A055,
         "Demographic.HRP.LCFS" = A003, 
         "HRP.Total.Hours.LCFS"= A220.x, 
         "HRP.Total.Wages.LCFS"= P004p.x,
         "HRP.Partner.Total.Hours.LCFS"=A220.y,
         "HRP.Partner.Total.Wages.LCFS" =P004p.y 
                                           )
```


```{r}

df_inflation<-read.csv("inflation.lookup.csv")

columns_to_match_rename <- c(
  "wages.inflation.22",
  "wages.inflation.23", 
  "wages.inflation.24"
)

# Loop through the columns and perform the matching and renaming
for (col_name in columns_to_match_rename) {
  LCFS_21_Work[[col_name]] <- df_inflation[[col_name]][match(LCFS_21_Work$Sample.Month, df_inflation$Sample.Month)]
}

```

# This will need adjusted for those on minimum wage
```{r}
LCFS_21_Work_Name<-LCFS_21_Work %>% 
  mutate(HRP.Total.Wages.22.LCFS=HRP.Total.Wages.LCFS*wages.inflation.22,
         HRP.Total.Wages.23.LCFS=HRP.Total.Wages.22.LCFS*wages.inflation.23,
         HRP.Total.Wages.24.LCFS=HRP.Total.Wages.23.LCFS*wages.inflation.24,
         HRP.Partner.Total.Wages.22.LCFS=HRP.Partner.Total.Wages.LCFS*wages.inflation.22,
         HRP.Partner.Total.Wages.23.LCFS=HRP.Partner.Total.Wages.22.LCFS*wages.inflation.23,
         HRP.Partner.Total.Wages.24.LCFS=HRP.Partner.Total.Wages.23.LCFS*wages.inflation.24
         )
LCFS_21_Work.21<-LCFS_21_Work_Name%>% mutate(Year="2020-21")
LCFS_21_Work.22<-LCFS_21_Work_Name%>% mutate(Year="2021-22P")
LCFS_21_Work.23<<-LCFS_21_Work_Name%>% mutate(Year="2022-23P")
LCFS_21_Work.24<<-LCFS_21_Work_Name%>% mutate(Year="2023-24P")

LCFS_21_Work <- rbind(LCFS_21_Work.21, LCFS_21_Work.22,LCFS_21_Work.23, LCFS_21_Work.24)

```

```{r}
LCFS_21_Work <- LCFS_21_Work %>% 
  mutate(HRP.Total.Wages.LCFS = case_when(Year == "2020-21" ~ HRP.Total.Wages.LCFS,
                                          Year == "2021-22P" ~ HRP.Total.Wages.22.LCFS,
                                          Year == "2022-23P" ~ HRP.Total.Wages.23.LCFS,
                                          Year == "2023-24P" ~ HRP.Total.Wages.23.LCFS)) %>% 

  mutate(HRP.Partner.Total.Wages.LCFS = case_when(Year == "2020-21" ~ HRP.Partner.Total.Wages.LCFS,
                                                  Year == "2021-22P" ~ HRP.Partner.Total.Wages.22.LCFS,
                                                  Year == "2022-23P" ~ HRP.Partner.Total.Wages.23.LCFS,
                                                  Year == "2023-24P" ~ HRP.Partner.Total.Wages.24.LCFS)) %>% 
            select(case,Year, Demographic.HRP.LCFS, HRP.Total.Hours.LCFS, HRP.Total.Wages.LCFS,
                   HRP.Partner.Total.Hours.LCFS,HRP.Partner.Total.Wages.LCFS)


saveRDS(LCFS_21_Work, "LCFS_21_Work.rds")
```

#HRP and Partner age
```{r}
#HRP Age and Partner Age
#from derived person level data
# dvper_22<-read.spss("lcfs_2021_dvper_ukanon202122.sav", #FY 2021-22
#                  use.value.labels = TRUE,
#                  use.missings = FALSE,
#                  to.data.frame = TRUE)
LCFS_21_dv_person<-readRDS("LCFS_21_dv_person.rds")

dvper_21<- LCFS_21_dv_person %>% 
  select(case, A002, BenUnPer, Person, P031, a006p, BenUnPer, a005p) 

dvper_2021_HRP<- dvper_21 %>% 
  select(case, A002, a005p) %>% filter(A002=="Person 1")

dvper_2021_Partner<- dvper_21 %>% 
  select(case, A002, a005p) %>%
  filter(A002=="Spouse/Cohabitee/Civil Partner")


dv_hrp_partner_2021<-dvper_2021_HRP %>% left_join(dvper_2021_Partner, by=c("case"="case")) %>% 
          
                      rename("Demographic.HRP.LCFS"=A002.x,
                             "Demographic.Relationship.LCFS"=A002.y,
                             "Demographic.HRP.Age.LCFS"=a005p.x,
                             "Demographic.HRP.Partner.Age.LCFS"=a005p.y)


dvper_20_21<-dv_hrp_partner_2021%>% mutate(Year="2020-21")
dvper_21_22<-dv_hrp_partner_2021%>% mutate(Year="2021-22P")
dvper_22_23<-dv_hrp_partner_2021%>% mutate(Year="2022-23P")
dvper_23_24<-dv_hrp_partner_2021%>% mutate(Year="2023-24P")


LCFS_21_Age <- rbind(dvper_20_21, dvper_21_22, dvper_22_23, dvper_23_24) %>% select( -Demographic.HRP.LCFS)

```

#HRP and Partner ethnicity
```{r}
df_ethnicity_HRP<-LCFS_21_dv_person %>% select(case, a012p) %>% filter(!is.na(a012p))

df_ethnicity_HRP_Partner<-LCFS_21_dv_person %>% select(case, a013p) %>% filter(!is.na(a013p))

LCFS_21_ethnicity<-df_ethnicity_HRP %>% left_join(df_ethnicity_HRP_Partner, by=c("case")) %>% rename("Demographic.ethnicity.HRP"=a012p,
                                                                                                "Demographic.ethnicity.HRP.Partner"=a013p)


```
